fix(macos): run multi-window probe as app target
