+++ .github/workflows/macos-lifecycle.yml
@@ -51,11 +51,12 @@ jobs:
           dart pub get
           dart run melos bootstrap
 
-      - name: Run minimal macOS multi-window regression test
+      - name: Run upstream macOS multi-window lifecycle probe
         working-directory: packages/bochki_schedule_app
         run: |
           set -o pipefail
-          flutter test integration_test/macos_minimal_multi_window_test.dart -d macos -r expanded 2>&1 | tee multi-window-lifecycle.log
+          flutter run -d macos -t lib/macos_upstream_multi_window_probe.dart 2>&1 | tee multi-window-lifecycle.log
+          rg -F '[macos-upstream-0.2.1] PASS upstream child lifecycle and message exchange' multi-window-lifecycle.log
           ! rg -n 'FlutterEngineOnVsync.*kInternalInconsistency|Unhandled platform error|WindowChannelException' multi-window-lifecycle.log
 
       - name: Upload macOS lifecycle log
